+++ src/main/java/com/elmakers/mine/bukkit/plugins/spells/builtin/ArrowSpell.java
@@ -1,6 +1,9 @@
 package com.elmakers.mine.bukkit.plugins.spells.builtin;
 
+import net.minecraft.server.EntityArrow;
+
 import org.bukkit.Material;
+import org.bukkit.craftbukkit.entity.CraftArrow;
 import org.bukkit.craftbukkit.entity.CraftPlayer;
 import org.bukkit.entity.Arrow;
 
@@ -8,20 +11,52 @@ import com.elmakers.mine.bukkit.plugins.spells.Spell;
 
 public class ArrowSpell extends Spell
 {
+    public ArrowSpell()
+    {
+        addVariant("arrowrain", Material.BOW, getCategory(), "Fire a volley of arrows", "20");
+    }
+    
 	@Override
 	public boolean onCast(String[] parameters)
 	{
+	    int arrowCount = 1;
+        if (parameters.length > 0)
+        {
+            try
+            {
+                arrowCount = Integer.parseInt(parameters[0]);
+            }
+            catch (NumberFormatException ex)
+            {
+                arrowCount = 1;
+            }
+        }
 		CraftPlayer cp = (CraftPlayer)player;
-		Arrow arrow = cp.shootArrow();
-		if (arrow == null)
+		
+		for (int ai = 0; ai < arrowCount; ai++)
 		{
-			castMessage(player, "Your arrow fizzled");
+		    Arrow arrow = cp.shootArrow();
+		    if (arrow == null)
+		    {
+		        sendMessage(player, "One of your arrows fizzled");
+		        return false;
+		    }
+		    if (ai != 0 && (arrow instanceof CraftArrow))
+		    {
+		        CraftArrow ca = (CraftArrow)arrow;
+		        EntityArrow ea = (EntityArrow)ca.getHandle();
+		        ea.setPosition
+		        (
+	                ea.locX + Math.random() * arrowCount - arrowCount / 2,
+	                ea.locY + Math.random() * arrowCount - arrowCount / 2,
+	                ea.locZ + Math.random() * arrowCount - arrowCount / 2
+		        );	                
+		    }
 		}
-		else
-		{
-			castMessage(player, "You fire a magical arrow");
-		}
-		return arrow != null;
+	
+		castMessage(player, "You fire some magical arrows");
+		
+		return true;
 	}
 
 	@Override