+++ src/main/java/com/elmakers/mine/bukkit/plugins/spells/builtin/FrostSpell.java
@@ -3,9 +3,12 @@ package com.elmakers.mine.bukkit.plugins.spells.builtin;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
+import org.bukkit.entity.Entity;
+import org.bukkit.entity.LivingEntity;
+import org.bukkit.entity.Player;
 
-import com.elmakers.mine.bukkit.persistence.dao.BlockList;
 import com.elmakers.mine.bukkit.plugins.spells.Spell;
+import com.elmakers.mine.bukkit.plugins.spells.Target;
 import com.elmakers.mine.bukkit.plugins.spells.utilities.PluginProperties;
 import com.elmakers.mine.bukkit.utilities.SimpleBlockAction;
 
@@ -55,17 +58,35 @@ public class FrostSpell extends Spell
 	@Override
 	public boolean onCast(String[] parameters)
 	{
-		Block target = getTargetBlock();
-		if (target == null)
-		{
-			castMessage(player, "No target");
-			return false;
-		}
-		if (defaultSearchDistance > 0 && getDistance(player, target) > defaultSearchDistance)
-		{
-			castMessage(player, "Can't frost that far away");
-			return false;
-		}
+	    Target target = getTarget();
+
+        if (target == null)
+        {
+            castMessage(player, "No target");
+            return false;
+        }
+        if (target.isEntity())
+        {
+            Entity targetEntity = target.getEntity();
+            if (targetEntity instanceof LivingEntity)
+            {
+                LivingEntity li = (LivingEntity)targetEntity;
+                if (li instanceof Player)
+                {
+                    li.damage(4);
+                }
+                else
+                {
+                    li.damage(10);
+                }
+            }
+        }
+
+        if (!target.hasTarget())
+        {
+            castMessage(player, "No target");
+            return false;
+        }
 		
 		int radius = defaultRadius;
 		if (parameters.length > 0)
@@ -88,7 +109,7 @@ public class FrostSpell extends Spell
 
         if (radius <= 1)
         {
-            action.perform(target);
+            action.perform(target.getBlock());
         }
         else
         {
@@ -99,62 +120,6 @@ public class FrostSpell extends Spell
         castMessage(player, "Frosted " + action.getBlocks().size() + " blocks");
         
         return true;
-	}
-	
-	public void frostBlock(int dx, int dy, int dz, Block centerPoint, int radius, BlockList frostedBlocks)
-	{
-		int x = centerPoint.getX() + dx - radius;
-		int y = centerPoint.getY() + dy - radius;
-		int z = centerPoint.getZ() + dz - radius;
-		Block block = player.getWorld().getBlockAt(x, y, z);
-		int depth = 0;
-		
-		if (block.getType() == Material.AIR)
-		{
-			while (depth < verticalSearchDistance && block.getType() == Material.AIR)
-			{
-				depth++;
-				block = block.getFace(BlockFace.DOWN);
-			}	
-		}
-		else
-		{
-			while (depth < verticalSearchDistance && block.getType() != Material.AIR)
-			{
-				depth++;
-				block = block.getFace(BlockFace.UP);
-			}
-			block = block.getFace(BlockFace.DOWN);
-		}
-
-		if (block.getType() == Material.AIR || block.getType() == Material.SNOW)
-		{
-			return;
-		}
-		Material material = Material.SNOW;
-		Block target = block;
-		if (block.getType() == Material.WATER || block.getType() == Material.STATIONARY_WATER)
-		{
-			material = Material.ICE;
-		}
-		else if (block.getType() == Material.LAVA)
-		{
-			material = Material.COBBLESTONE;
-		}
-		else if (block.getType() == Material.STATIONARY_LAVA)
-		{
-			material = Material.OBSIDIAN;
-		}
-		else if (block.getType() == Material.FIRE)
-		{
-			material = Material.AIR;
-		}
-		else
-		{
-			target = target.getFace(BlockFace.UP);
-		}
-		frostedBlocks.add(target);
-		target.setType(material);
 	}
 
 	public int checkPosition(int x, int z, int R)
@@ -171,7 +136,7 @@ public class FrostSpell extends Spell
 	@Override
 	public String getCategory()
 	{
-		return "nature";
+		return "combat";
 	}
 
 	@Override