+++ src/main/java/com/elmakers/mine/bukkit/plugins/spells/builtin/TransmuteSpell.java
@@ -84,7 +84,7 @@ public class TransmuteSpell extends Spell
 	@Override
 	public String getDescription()
 	{
-		return "Modify your last construction";
+		return "Modify your last spell";
 	}
 
 	@Override